requirements.txt
emmet-core==0.49.1
keras==2.12.0
matplotlib==3.7.1
mp-api==0.30.10
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
tensorflow==2.12.0